python-api/requirements.txt
fastapi
uvicorn
python-Levenshtein
mysql-connector-python
python-dotenv
